create some components
